// src/types/api.ts
export interface SignupResponse {
  success: boolean;
  message: string;
}